--- backend/src/main/java/org/cathori/backend/alert/api/dto/NotificationListResponse.java
@@ -1,9 +1,0 @@
-package org.cathori.backend.alert.api.dto;
-
-import java.util.List;
-
-public record NotificationListResponse(
-        List<NotificationItem> alerts,
-        Long nextCursor,
-        boolean hasNext
-) {}
