--- backend/src/main/java/org/cathori/backend/alert/application/notification/NotificationQueryPort.java
@@ -1,7 +1,0 @@
-package org.cathori.backend.alert.application.notification;
-
-import java.util.List;
-
-public interface NotificationQueryPort {
-    List<NotificationRow> findByUserIdWithCursor(Long userId, Long cursor, int limit);
-}
