--- backend/src/main/java/org/cathori/backend/alert/application/push/UserToken.java
@@ -1,4 +1,0 @@
-package org.cathori.backend.alert.application.push;
-
-public record UserToken(Long userId, String token) {
-}
