+++ backend/src/main/java/org/cathori/backend/notification/api/dto/NotificationItem.java
@@ -1,0 +1,14 @@
+package org.cathori.backend.notification.api.dto;
+
+import java.time.LocalDate;
+import java.time.OffsetDateTime;
+
+public record NotificationItem(
+        Long alertHistoryId,
+        Long noticeId,
+        String title,
+        String matchedTag,
+        LocalDate deadlineAt,
+        boolean isRead,
+        OffsetDateTime createdAt
+) {}
