+++ backend/src/main/java/org/cathori/backend/notification/api/dto/NotificationListResponse.java
@@ -1,0 +1,9 @@
+package org.cathori.backend.notification.api.dto;
+
+import java.util.List;
+
+public record NotificationListResponse(
+        List<NotificationItem> alerts,
+        Long nextCursor,
+        boolean hasNext
+) {}
