+++ backend/src/main/java/org/cathori/backend/notification/application/inbox/NotificationQueryPort.java
@@ -1,0 +1,7 @@
+package org.cathori.backend.notification.application.inbox;
+
+import java.util.List;
+
+public interface NotificationQueryPort {
+    List<NotificationRow> findByUserIdWithCursor(Long userId, Long cursor, int limit);
+}
