+++ backend/src/main/java/org/cathori/backend/notification/application/push/UserFcmResult.java
@@ -1,0 +1,4 @@
+package org.cathori.backend.notification.application.push;
+
+public record UserFcmResult(Long userId, boolean success) {
+}
