+++ backend/src/main/java/org/cathori/backend/notification/application/push/UserToken.java
@@ -1,0 +1,4 @@
+package org.cathori.backend.notification.application.push;
+
+public record UserToken(Long userId, String token) {
+}
